requirements.txt
requests
pandas
openpyxl
matplotlib
python-dotenv
jsonpickle
seaborn
